website/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Babel==2.10.1
beautifulsoup4==4.9.3
bidict==0.22.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Deprecated==1.2.13
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.0
Flask==2.1.1
Flask-Assets==2.0
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-Login==0.6.0
flask-mongoengine==1.0.0
Flask-SocketIO==5.1.2
Flask-WTF==1.0.1
flower==1.0.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
greenletio==0.9.0
gunicorn==20.1.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.0
Jinja2==3.1.1
jsmin==3.0.1
kombu==5.2.4
logger-slg==0.3.5
MarkupSafe==2.1.1
mongoengine==0.24.1
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.8
python-dotenv==0.20.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
selenium==3.141.0
six==1.16.0
slg-utilities==0.4.16
soupsieve==2.2.1
tabulate==0.8.9
tornado==6.1
typing-extensions==4.1.1
urllib3==1.26.5
vine==5.0.0
visitor==0.1.3
watchdog==2.1.7
wcwidth==0.2.5
webassets==2.0
webdriver-manager==3.4.2
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
